+++ src/file/writer.rs
@@ -12,7 +12,7 @@ use super::File;
 /// [Write] implementation for [File]s
 pub struct Writer<'r, 'f> {
     file: &'r mut File<'f>,
-    pos: usize,
+    pos: u64,
 }
 
 impl<'f> File<'f> {
@@ -38,7 +38,7 @@ impl<'r, 'f> Writer<'r, 'f> {
         let write_start = self.pos;
         let write_end = write_start + ext_len;
         if let Some((existing_start, existing_ext)) = self.file.extent_for_byte_mut(write_end) {
-            let right = existing_ext.split_at(write_end - existing_start);
+            let right = existing_ext.split_at((write_end - existing_start) as usize);
             self.file.extents.insert(write_end, right);
         }
         if let Some((existing_start, existing_ext)) = self.file.extent_for_byte_mut(self.pos) {
@@ -46,9 +46,9 @@ impl<'r, 'f> Writer<'r, 'f> {
             // shrink this extent to end where the overlap is
             let split_idx = write_start - existing_start;
             let right_split_idx = write_end - split_idx;
-            let mut right = existing_ext.split_at(split_idx);
+            let mut right = existing_ext.split_at(split_idx as usize);
             if right_split_idx < right.len() {
-                right.split_at(right_split_idx);
+                right.split_at(right_split_idx as usize);
                 let right_start = write_end;
                 self.file.extents.insert(right_start, right);
             }
@@ -73,16 +73,16 @@ impl<'r, 'f> Seek for Writer<'r, 'f> {
     fn seek(&mut self, seek: SeekFrom) -> Result<u64> {
         let (base_pos, offset) = match seek {
             SeekFrom::Start(n) => {
-                self.pos = n as usize;
+                self.pos = n;
                 return Ok(n);
             }
-            SeekFrom::End(n) => (self.file.len() as u64, n),
-            SeekFrom::Current(n) => (self.pos as u64, n),
+            SeekFrom::End(n) => (self.file.len(), n),
+            SeekFrom::Current(n) => (self.pos, n),
         };
         match base_pos.checked_add_signed(offset) {
             Some(n) => {
-                self.pos = n as usize;
-                Ok(self.pos as u64)
+                self.pos = n;
+                Ok(self.pos)
             }
             None => Err(Error::new(
                 ErrorKind::InvalidInput,
@@ -122,7 +122,7 @@ mod tests {
             &f.extents,
             &BTreeMap::from([
                 (0, "Lorem ".into()),
-                ("Lorem ".len(), "ipsum dolor sit amet".into()),
+                ("Lorem ".len() as u64, "ipsum dolor sit amet".into()),
             ]),
         );
     }
@@ -146,8 +146,8 @@ mod tests {
             &f.extents,
             &BTreeMap::from([
                 (0, "Lorem ".into()),
-                ("Lorem ".len(), "ipsum".into()),
-                ("Lorem ipsum".len(), " dolor sit amet".into()),
+                ("Lorem ".len() as u64, "ipsum".into()),
+                ("Lorem ipsum".len() as u64, " dolor sit amet".into()),
             ]),
         );
     }
